fix: satisfy KSeF test coding standard
