+++ tests/lib/KSeF/ConfigHelper.php
@@ -1,0 +1,11 @@
+<?php
+
+if (!class_exists('ConfigHelper')) {
+    class ConfigHelper
+    {
+        public static function checkConfig()
+        {
+            return false;
+        }
+    }
+}
