+++ tests/lib/KSeF/LMS.php
@@ -1,0 +1,9 @@
+<?php
+
+if (!class_exists('LMS')) {
+    class LMS
+    {
+        const SOFTWARE_NAME = 'LMS';
+        const SOFTWARE_VERSION = 'test';
+    }
+}
