+++ tests/lib/KSeF/Localisation.php
@@ -1,0 +1,11 @@
+<?php
+
+if (!class_exists('Localisation')) {
+    class Localisation
+    {
+        public static function getCurrentCurrency()
+        {
+            return 'PLN';
+        }
+    }
+}
